Improve HAMT perf + add performance specs

+++ src/main/scala/data_structures/PerfUtils.scala
@@ -7,7 +7,7 @@ object PerfUtils {
   def time[A](id: String)(block: => A): A = {
     val startTime = System.nanoTime
     val result = block
-    val elapsedMicro = Math.round((System.nanoTime() - startTime) / 1000)
+    val elapsedMicro = Math.round((System.nanoTime() - startTime).toFloat / 1000)
     val elapsedStr = if (elapsedMicro > 5000) s"${elapsedMicro / 1000}ms" else s"${elapsedMicro}us"
     println(s"$id took $elapsedStr")
     result
